add RecipeGrid component
